npu: MoE checks out, and both of its known defects now have numbers
The layer itself is right: moe_out lands at 0.25-0.57x budget at decode bs=16
and an 8192 prefill chunk, on real TP16 shapes with the production dispatcher
and runner. The only quantity over budget was router_logits, which is defect (2).

Defect (1), DeepEP dropping swiglu_limit, is real but GLM's factory recipe never
reaches it: --moe-a2a-backend none routes to AscendTPDispatcher, whose
activation already carries the limit. Only DeepSeek-V4's deepep recipe hits it.
Same shape as OP-3 -- a real defect on a path this model does not take -- and
worth stating both halves. It is fixed anyway, by passing the do_limit/limit
arguments swiglu_quant has always had. On real layer-3 gmm1 output that moves it
from 2.85x budget to 0.35x, the same number the factory path gets, and layer 40
is untouched because nothing there exceeds the limit at all.

That fix changes DeepSeek-V4's numerics, since deepep is its factory recipe. The
direction is right -- its shared experts already clamp, so the layer is
internally inconsistent today -- but it has not been regressed against DSv4's
GPQA baseline, and the plan says so.

Defect (2), the bf16 router GEMM, is fixed here. Reading the branches settles
what looked like a shared-code risk: every CUDA path in MoEGate.forward goes out
of its way for an fp32 result, one of them with a comment explaining the
detour, and `elif not _is_cuda` is the only branch that stops at bf16. Making it
fp32 is convergence, not a behaviour change.

It matters more than its magnitude suggests because the loss lands on a top-k,
so it surfaces as a discretely wrong expert rather than a small error: 5.65% of
tokens at layer 3 picked a different expert set, the worst of them 34% off,
while the rel-L2 for that same case sat inside budget. Overlap caught what the
norm could not. fp32 restores overlap exactly to the floor for +36us at decode
and +383us on an 8192 chunk.

Also: moe_router_dtype, which this model's config sets, has no consumer
anywhere in the repo.

+++ python/sglang/srt/models/deepseek_v2.py
@@ -565,7 +565,19 @@ class MoEGate(nn.Module):
             elif _use_aiter:
                 logits = aiter_dsv3_router_gemm(hidden_states, self.weight)
             elif not _is_cuda:
-                logits = F.linear(hidden_states, self.weight, None)
+                # fp32, like every CUDA branch above: they each go out of their
+                # way for an fp32 result, and this one is the only path that
+                # stops at bf16. It matters more than the magnitude suggests --
+                # the loss lands on a top-k, so it shows up as a discrete wrong
+                # expert rather than a small error. Measured on Ascend at layer
+                # 3 / 8192 tokens: bf16 logits give 0.9929 top-8 set overlap,
+                # 5.65% of tokens picking a different expert set and the worst
+                # of them 34% off; fp32 restores the overlap exactly to the
+                # bf16-input noise floor. Costs +36us at decode bs=16 and
+                # +383us on an 8192 prefill chunk.
+                logits = F.linear(
+                    hidden_states.float(), self.weight.float(), None
+                )
             else:
                 # cuBLAS bf16 x bf16 -> fp32 GEMM (torch.mm's out_dtype kwarg is CUDA-only)
                 from sglang.kernels.ops.attention.dsv4 import linear_bf16_fp32
